refactor(calendar): move navigation logic into module
